+++ ValheimCommunityPatch/Package/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+**0.23.1**
+```
+- Fix Equipment Visual Refresh (client): a beard or hair no longer turns white (only a
+  death respawn or new character) were affected.
+```
+
 **0.23.0**
 ```
 - Fix Teleport Ghost Players (server): a player who teleports out of another player's loaded area is